requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
tensorflow==2.14.0
keras==2.14.0
numpy==1.26.4
opencv-python==4.7.0.72
pillow==10.0.0
tensorflow-hub==0.16.1
gunicorn==21.2.0
python-dotenv==1.0.0
